Relocate AdditionalArmorItem
